Rubocop

We must be aware that once we merge this we need to correct all errors.

See merge request !70

+++ app/controllers/access_policies_controller.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 class AccessPoliciesController < ApplicationController
   def new
     @resource = Resource.find_by(id: params[:resource_id])
@@ -6,7 +7,7 @@ class AccessPoliciesController < ApplicationController
       @access_policy = AccessPolicy.new
       set_other_fields
     else
-      flash[:alert] = t("resource_not_found")
+      flash[:alert] = t('resource_not_found')
       redirect_to resources_path
     end
   end
@@ -15,9 +16,7 @@ class AccessPoliciesController < ApplicationController
     @access_policy = AccessPolicy.new(access_policy_params)
     @resource = Resource.find_by(id: @access_policy.resource_id)
 
-    if !@resource
-      flash[:alert] = t("resource_not_found")
-    end
+    flash[:alert] = t('resource_not_found') unless @resource
 
     set_other_fields
 
@@ -31,9 +30,7 @@ class AccessPoliciesController < ApplicationController
   def destroy
     access_policy = AccessPolicy.find_by(id: params[:id])
 
-    if access_policy
-      access_policy.destroy
-    end
+    access_policy.destroy if access_policy
 
     redirect_to new_access_policy_path(resource_id: params[:resource_id])
   end
@@ -45,17 +42,24 @@ class AccessPoliciesController < ApplicationController
     @groups = Group.all
     @access_methods = AccessMethod.all
 
-    @user_access_policies = {}
-    User.joins(access_policies: :resource).includes(access_policies: :access_method)
-        .where(resources: {id: @resource.id}).each do |user|
-      @user_access_policies[user.email] = user.access_policies.where(resource_id: @resource.id)
-    end
+    set_user_access_policies
+    set_group_access_policies
+  end
 
+  def set_group_access_policies
     @group_access_policies = {}
-    Group.joins(access_policies: :resource).includes(access_policies: :access_method)
-        .where(resources: {id: @resource.id}).each do |group|
-      @group_access_policies[group.name] = group.access_policies.where(resource_id: @resource.id)
-    end
+    Group.joins(access_policies: :resource).includes(access_policies: :access_method).
+      where(resources: { id: @resource.id }).each do |group|
+        @group_access_policies[group.name] = group.access_policies.where(resource_id: @resource.id)
+      end
+  end
+
+  def set_user_access_policies
+    @user_access_policies = {}
+    User.joins(access_policies: :resource).includes(access_policies: :access_method).
+      where(resources: { id: @resource.id }).each do |user|
+        @user_access_policies[user.email] = user.access_policies.where(resource_id: @resource.id)
+      end
   end
 
   def access_policy_params
